aspeed: Add wks file to build a wic image for eMMC
Add a wks file that can be used to build a wic image for a eMMC chip.

The first partition holds the fitImage file. Code update can then
convert this file to a symlink fitImage -> kernel-<versionid> so
that the BMC loads the updated kernel file.

The remaining is a set of ext4 filesystems that can be used to
created device mappers. See:
https://gerrit.openbmc-project.xyz/c/openbmc/meta-phosphor/+/32610

Change-Id: I37759760ac7baa42ab9d202f0926d4a31d0070ac
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>

squash with wic

Change-Id: I4c6f41fdb1b21ac6a4e7f4e5497a9ac966ffd31e