obmc-bsp-common: Build default eMMC images
OpenBMC only supported 2 layouts: static and UBI. With the
introduction of eMMC, need to change the logic so that
static images are only built if UBI or eMMC are not
specified. Add a default eMMC image to be built.

Tested: Adding this line to a machine conf only built the
        emmc tarball:
        require conf/distro/include/phosphor-mmc.inc

Change-Id: Ic626009981808519e31a36cbb535ced74a362858
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>